Add files via upload
A Computer Vision to classify Dog and Cat image. Model created using Keras sequential model, It is a Convolution Neural Network (CNN) with 2 layers. The model is trained on 30 epochs and has an accuracy of 84%. Also created a simple GUI in Django to interact with the model. 
Dataset can be found at - https://www.kaggle.com/chetankv/dogs-cats-images